provide pytest plugin to simplify testing in sub-projects
