--- pytest.ini
@@ -1,6 +1,0 @@
-[pytest]
-# NOTE: changing default location of pytest_cache because the uvicorn file watcher somehow triggered too many reloads
-cache_dir = ../.pytest_cache 
-testpaths =
-    tests
-asyncio_mode = auto
